Fix the math of the number of subregions
The hregion algorithm works on the basis of a statically
allocated array which is described in terms of the number of
layers and the number of subregions.

A layer corresponds to a displayframe that be divided into a number
of subregions. It is asserted that when adding layers with increasing
z-order that only a maximum of two more subregions will be created.

The net effect is that the worst case number of subregions is
(max. layers * 2) - 1.

layers     1 2 3 4---4 3 2 1
subregions | | | |   | | | |
            1 2 3  4  5 6 7

In the future we may have to consider layers which have multiple
rectangles at the same z.

Also updated the constants names to better reflect their role.

Signed-off-by: Tony Lofthouse <a0741364@ti.com>